requirements.txt
# This file lists the Python packages required to run the AnnoForge application.
# Install them using the following command:
# pip install -r requirements.txt

Pillow>=11.3.0
piexif>=1.1.3         # For reading/writing metadata in the EXIF Editor plugin.

torch>=2.7.1
transformers==4.49.0  # Hugging Face library to use the Florence-2 model.
numpy>=2.2.6
timm>=1.0.19

customtkinter
pyperclip
tkinterdnd2-universal